10/18/16 - Millburn
I really like a lot of this but in the beginning you seemed to set it up for a description of the aggressive practices that keep the town white, but beyond the police hassling, there wasn't much of that discussed. I like what follows alot....you might want to redo the intro to match it? And were you planning on working on this a little more? I kind of like the trailing off of "some hope" at the end, but it didn't seem like it really ended there.  But this is bold, important stuff that no one talks about... and it's just so, so NJ.