10/18/16 - Newark
Here you have the story of the aggressive maltreatment of the black community that you kind of started in the Millburn piece. Perhaps all you need to do is add a line at the end of that first paragraph in the Millburn article, in parens, along the lines of (For more on this, see the article on Newark) or something like that.