10/27/16 - South and East Asia
Tempting to compare the Asian influx to the earlier Jewish one; both came with strong cultural identities that found fertile ground in the US but also brought on conflicts of assimilation. The Jews came from poverty in most cases, however, while the Asians come from higher classes, it appears. Both the Jews and Asians started filling up the orchestras and the top academic slots...